requirements.txt
Django==4.2.6
django-ckeditor==6.7.0
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==10.1.0
python-dotenv==1.0.0
